% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{dissertation_epub}
\alias{dissertation_epub}
\title{Creates an R Markdown epub dissertation document}
\usage{
dissertation_epub(...)
}
\arguments{
\item{...}{other arguments to bookdown::epub_book}
}
\value{
A ebook version of the thesis
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a epub version of the dissertation.
}
\examples{
\dontrun{
 output: eagledown::dissertation_epub
}
}
