% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{dissertation_word}
\alias{dissertation_word}
\title{Creates an R Markdown Word dissertation document}
\usage{
dissertation_word(...)
}
\arguments{
\item{...}{other arguments to  bookdown::word_document2}
}
\value{
A Word Document based on (not currently)
the American University Word dissertation template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the dissertation.
}
\examples{
\dontrun{
 output: eagledown::dissertation_word
}
}
