% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagledown.R
\docType{package}
\name{eagledown}
\alias{eagledown}
\alias{eagledown-package}
\title{eagledown: A package for creating a PhD dissertations
using R Markdown based on the American University
LaTeX dissertation template}
\description{
eagledown: A package for creating a PhD dissertations
using R Markdown based on the American University
LaTeX dissertation template
}
